+++ boot/src/x64.rs
@@ -28,26 +28,3 @@ pub fn hlt() {
     }
 }
 
-pub fn set_segment_registers(code_segment_selector: &memory::segment::Selector, data_segment_selector: &memory::segment::Selector) {
-    let code_segment_selector: u16 = code_segment_selector.clone().into();
-    let data_segment_selector: u16 = data_segment_selector.clone().into();
-    unsafe {
-        asm!(
-            "mov ds, {data_segment_selector:x}",
-            "mov es, {data_segment_selector:x}",
-            "mov fs, {data_segment_selector:x}",
-            "mov gs, {data_segment_selector:x}",
-            "mov ss, {data_segment_selector:x}",
-            "movzx {extended_code_segment_selector}, {code_segment_selector:x}",
-            "lea {destination}, [rip + 0f]",
-            "push {extended_code_segment_selector}",
-            "push {destination}",
-            "retfq",
-            "0:",
-            code_segment_selector = in(reg) code_segment_selector,
-            data_segment_selector = in(reg) data_segment_selector,
-            extended_code_segment_selector = out(reg) _,
-            destination = out(reg) _,
-        );
-    }
-}
