+++ kernel/src/memory/segment/descriptor/table.rs
@@ -11,6 +11,7 @@ use {
     core::{
         fmt,
         mem,
+        ops::Range,
         slice,
     },
     super::{
