+++ kernel/src/memory/segment/selector.rs
@@ -72,6 +72,10 @@ impl Selector {
         fs.into()
     }
 
+    pub fn get_index(&self) -> u16 {
+        self.index()
+    }
+
     pub fn get_rpl(&self) -> u8 {
         self.rpl()
     }
