+++ kernel/src/x64.rs
@@ -17,6 +17,8 @@ pub use {
     rflags::Rflags,
 };
 
+use crate::memory;
+
 /// # Halt
 /// ## References
 /// * [Intel 64 and IA-32 Architectures Software Developer's Manual December 2023](https://www.intel.com/content/www/us/en/developer/articles/technical/intel-sdm.html) Vol.2A 3-489
@@ -24,6 +26,30 @@ pub use {
 pub fn hlt() {
     unsafe {
         asm!("hlt");
+    }
+}
+
+pub fn set_segment_registers(code_segment_selector: &memory::segment::Selector, data_segment_selector: &memory::segment::Selector) {
+    let code_segment_selector: u16 = code_segment_selector.clone().into();
+    let data_segment_selector: u16 = data_segment_selector.clone().into();
+    unsafe {
+        asm!(
+            "mov ds, {data_segment_selector:x}",
+            "mov es, {data_segment_selector:x}",
+            "mov fs, {data_segment_selector:x}",
+            "mov gs, {data_segment_selector:x}",
+            "mov ss, {data_segment_selector:x}",
+            "movzx {extended_code_segment_selector}, {code_segment_selector:x}",
+            "lea {destination}, [rip + 0f]",
+            "push {extended_code_segment_selector}",
+            "push {destination}",
+            "retfq",
+            "0:",
+            code_segment_selector = in(reg) code_segment_selector,
+            data_segment_selector = in(reg) data_segment_selector,
+            extended_code_segment_selector = out(reg) _,
+            destination = out(reg) _,
+        );
     }
 }
